Infra - new -h command for CL_tool upload

+++ lib/infra/helpers.rb
@@ -24,11 +24,13 @@ module Infra
         args
       end
       list.each do |name|
-        FileUtils::cd(name) do
-          files = Dir.glob('*')
-          files.each do |file|
-            FileUtils.rm_rf(file, :secure => true)
-          end
+        if (name.directory?) then
+         FileUtils::cd(name) do
+            files = Dir.glob('*')
+            files.each do |file|
+              FileUtils.rm_rf(file, :secure => true)
+            end
+         end
         end
       end
     end
@@ -104,9 +106,7 @@ module Infra
       fail ArgumentError.new("Error in Upload_data_with_cl helper. Please define login either as parameter LOGIN in params.json or as :login option") if login.nil? || login.empty?
       fail ArgumentError.new("Error in Upload_data_with_cl helper. Please define login either as parameter PASSWORD in params.json or as :password option") if password.nil? || password.empty?
 
-
-      puts "#{get('CLTOOL_EXE')} -u#{login} -p#{password} #{optional("-h",host)} #{script_path}"
-      #run_shell("#{get('CLTOOL_EXE')} -u#{login} -p#{password} #{optional("-h",host)} #{script_path}")
+      run_shell("#{get('CLTOOL_EXE')} -u#{login} -p#{password} #{optional("-h",host)} #{script_path}")
     end
 
     def run_clover_graph(graph, options={})
